avoid arithmetic overflows

+++ ThrEd4/trace.cpp
@@ -318,7 +318,7 @@ void trace::trdif() {
 	StateMap.reset(StateFlag::HIDMAP);
 	trace::untrace();
 	if (BitmapHeight * BitmapWidth) {
-		std::vector<unsigned> differenceBitmap(BitmapHeight * BitmapWidth);
+		std::vector<unsigned> differenceBitmap(gsl::narrow<size_t>(BitmapHeight) * BitmapWidth);
 
 		auto colorSumMaximum = 0u;
 		auto colorSumMinimum = 0xffffffffu;
@@ -467,14 +467,14 @@ void trace::tracedg() {
 			}
 			else {
 				if (flag) {
-					TracedEdges->set(pixelIndex - 1);
+					TracedEdges->set(gsl::narrow<size_t>(pixelIndex) - 1);
 					flag = false;
 				}
 			}
 			pixelIndex++;
 		}
 		if (flag) {
-			TracedEdges->set(pixelIndex - 1);
+			TracedEdges->set(gsl::narrow<size_t>(pixelIndex) - 1);
 		}
 	}
 	for (auto iWidth = 0u; iWidth < BitmapWidth; iWidth++) {
@@ -489,14 +489,14 @@ void trace::tracedg() {
 			}
 			else {
 				if (flag) {
-					TracedEdges->set(pixelIndex - BitmapWidth);
+					TracedEdges->set(gsl::narrow<size_t>(pixelIndex) - BitmapWidth);
 					flag = false;
 				}
 			}
 			pixelIndex += BitmapWidth;
 		}
 		if (flag) {
-			TracedEdges->set(pixelIndex - BitmapWidth);
+			TracedEdges->set(gsl::narrow<size_t>(pixelIndex) - BitmapWidth);
 		}
 	}
 	for (auto iPixel = 0u; iPixel < BitmapWidth * BitmapHeight; iPixel++) {
